+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Parse payload content (JSON/UTF8-String/other) only once. Before it was done on every display update.
+
 ## [0.16.1] - 2022-03-23
 
 ### Added